requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cytoolz==0.11.2
dnspython==2.2.1
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
etherscan-python==2.1.0
fastapi==0.75.1
frozenlist==1.3.0
h11==0.12.0
hexbytes==0.2.2
httpcore==0.14.7
httpx==0.22.0
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.22.3
pandas==1.4.2
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.5.1
protobuf==3.20.0
pycryptodome==3.14.1
pydantic==1.9.0
pymongo==4.1.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==303
requests==2.27.1
rfc3986==1.5.0
rlp==2.0.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
toolz==0.11.2
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
varint==1.0.2
web3==5.28.0
websockets==9.1
yarl==1.7.2
